splite file
